pyproject.toml
[project]
name = "statskita"
version = "0.1.0"
description = "Python toolkit for Indonesian official microdata (SAKERNAS)"
readme = "README.md"
authors = [
    { name = "Okky Mabruri", email = "okkymbrur@gmail.com" }
]
license = { text = "MIT" }
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Topic :: Sociology"
]
keywords = ["indonesia", "statistics", "sakernas", "survey", "microdata", "labor", "employment"]
dependencies = [
    "polars>=0.20.0",
    "pyreadstat>=1.2.0",
    "samplics>=0.4.0",
    "pyarrow>=10.0.0",
    "xlsxwriter>=3.2.9",
    "requests>=2.28.0",
    "pydantic>=2.0.0",
    "typing-extensions>=4.5.0",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.0",
    "dbfrs>=0.1.5",
    "pandas>=2.3.2",
]

[project.optional-dependencies]
viz = [
    "matplotlib>=3.6.0",
    "seaborn>=0.12.0",
]
notebook = [
    "ipykernel>=6.0.0",
    "matplotlib>=3.6.0",
    "seaborn>=0.12.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "pre-commit>=3.0.0"
]
cli = [
    "typer>=0.9.0",
    "rich>=13.0.0"
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
    "mkdocstrings[python]>=0.22.0"
]

# CLI disabled for now
# [project.scripts]
# statskita = "statskita.cli:main"

[project.urls]
Homepage = "https://github.com/okkymabruri/statskita"
Documentation = "https://statskita.readthedocs.io"
Repository = "https://github.com/okkymabruri/statskita"
Issues = "https://github.com/okkymabruri/statskita/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "F",  # pyflakes
    "I",  # isort
    "W",  # pycodestyle warnings
]
ignore = ["E501"]  # line too long

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[dependency-groups]
dev = [
    "openpyxl>=3.1.5",
    "pytest>=8.4.2",
    "ruff>=0.13.2",
    "twine>=6.2.0",
]
